api/requirements.txt
Flask==1.0.2
tensorflow==1.11.0
Pillow==5.3.0